finish tidy up extention and types
