+++ services/madoc-ts/src/frontend/admin/pages/topics/tags/ListCanvasTags.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { EnrichmentResourceResponse, EntityTagSnippet } from '../../../../../extensions/enrichment/authority/types';
+import { EnrichmentResource, EntityTagSnippet } from '../../../../../extensions/enrichment/types';
 import { useData } from '../../../../shared/hooks/use-data';
 import { serverRendererFor } from '../../../../shared/plugins/external/server-renderer-for';
 import { ParseResourceTags } from '../../../../site/hooks/canvas-menu/use-get-tags';
@@ -18,7 +18,7 @@ import { AddTopicButton } from '../../../../site/features/AddTopicButton';
 import { Button, SmallButton } from '../../../../shared/navigation/Button';
 
 export function ListCanvasTags() {
-  const { data, isError, refetch } = useData<EnrichmentResourceResponse>(ListCanvasTags);
+  const { data, isError, refetch } = useData<EnrichmentResource>(ListCanvasTags);
   const ResourceTags = ParseResourceTags(data?.entity_tags);
   const api = useApi();
   const { t } = useTranslation();
@@ -116,6 +116,6 @@ serverRendererFor(ListCanvasTags, {
     return ['canvas-resource', { id: params.id }];
   },
   getData: async (key: string, vars, api) => {
-    return await api.enrichment.getSiteEnrichmentResource(`urn:madoc:canvas:${vars.id}`);
+    return await api.getSiteEnrichmentResource(`urn:madoc:canvas:${vars.id}`);
   },
 });
