+++ services/madoc-ts/src/frontend/admin/pages/topics/tags/ListManifestTags.tsx
@@ -1,5 +1,5 @@
 import { useData } from '../../../../shared/hooks/use-data';
-import { EnrichmentResourceResponse, EntityTagSnippet } from '../../../../../extensions/enrichment/authority/types';
+import { EnrichmentResource, EntityTagSnippet } from '../../../../../extensions/enrichment/types';
 import { ParseResourceTags } from '../../../../site/hooks/canvas-menu/use-get-tags';
 import { TableActions, TableContainer, TableRow, TableRowLabel } from '../../../../shared/layout/Table';
 import { TagPill } from '../../../../site/hooks/canvas-menu/tagging-panel';
@@ -18,7 +18,7 @@ import { Accordion } from '../../../../shared/atoms/Accordion';
 import { useTranslation } from 'react-i18next';
 
 export function ListManifestTags() {
-  const { data, isError, refetch } = useData<EnrichmentResourceResponse>(ListManifestTags);
+  const { data, isError, refetch } = useData<EnrichmentResource>(ListManifestTags);
   const ResourceTags = ParseResourceTags(data?.entity_tags);
   const api = useApi();
   const { t } = useTranslation();
@@ -115,6 +115,6 @@ serverRendererFor(ListManifestTags, {
     return ['manifest-resource', { id: params.id }];
   },
   getData: async (key: string, vars, api) => {
-    return await api.enrichment.getSiteEnrichmentResource(`urn:madoc:manifest:${vars.id}`);
+    return await api.getSiteEnrichmentResource(`urn:madoc:manifest:${vars.id}`);
   },
 });
